+++ src/movies/movies.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { TmdbApiModule } from './movie-api/tmdb-api/tmdb-api.module';
+import { TmdbApiModule } from './movies-api/tmdb-api/tmdb-api.module';
 import { MoviesController } from './movies.controller';
 import { MoviesService } from './movies.service';
